requirements.txt
flask==2.0.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
werkzeug==2.0.1
numpy==1.24.0 